+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2001.1.2]
+
+### Fixed
+* Fixed occasional client issue with client player being null when a creative tab rebuild done
+  * Led to player login to dedicated server sometimes failing
+
 ## [2001.1.1]
 
 ### Changed